requirements.txt
torch==2.5.1+cu118
torchvision==0.20.1+cu118
Pillow==10.3.0
matplotlib==3.9.0
